pyproject.toml
[tool.poetry]
name = "flask schedule"
version = "0.1.0"
description = ""
authors = ["Diana Perugia Baiadori <diana.baiadori@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-restful = "^0.3.9"
flask-apispec = "^0.11.4"
google-api-python-client = "^2.65.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.0"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
psycopg2-binary = "^2.9.5"
flask-sqlalchemy = "^3.0.2"


[tool.poetry.group.dev.dependencies]
pylint-flask = "^0.6"
pylint-flask-sqlalchemy = "^0.2.0"
pylint = "^2.15.5"
flake8 = "^5.0.4"
pytest = "^7.2.0"
coverage = "^6.5.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
